refactor(grid overlay): the types were moved to the component itself
