+++ src/components/Debugger/components/GridOverlay/index.ts
@@ -1,0 +1,1 @@
+export { default as GridOverlay } from './GridOverlay';
